+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -3,8 +3,8 @@ package techcourse.myblog.service;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.domain.exception.UserArgumentException;
-import techcourse.myblog.dto.UserDto;
-import techcourse.myblog.dto.UserPublicInfoDto;
+import techcourse.myblog.service.dto.UserDto;
+import techcourse.myblog.service.dto.UserPublicInfoDto;
 import techcourse.myblog.repository.ArticleRepository;
 import techcourse.myblog.repository.UserRepository;
 import techcourse.myblog.service.exception.NotFoundUserException;
