+++ src/main/java/techcourse/myblog/service/dto/LogInInfoDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
 public class LogInInfoDto {
     private String email;
