+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -3,8 +3,8 @@ package techcourse.myblog.web;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.dto.ArticleDto;
-import techcourse.myblog.dto.UserPublicInfoDto;
+import techcourse.myblog.service.dto.ArticleDto;
+import techcourse.myblog.service.dto.UserPublicInfoDto;
 import techcourse.myblog.service.ArticleService;
 import techcourse.myblog.service.UserService;
 import techcourse.myblog.web.exception.NotLoggedInException;
