+++ src/main/java/techcourse/myblog/web/MyPageController.java
@@ -5,7 +5,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import techcourse.myblog.dto.UserPublicInfoDto;
+import techcourse.myblog.service.dto.UserPublicInfoDto;
 import techcourse.myblog.service.UserService;
 
 import javax.servlet.http.HttpServletRequest;
