+++ src/main/java/techcourse/myblog/web/advice/ControllerExceptionAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import techcourse.myblog.dto.UserPublicInfoDto;
+import techcourse.myblog.service.dto.UserPublicInfoDto;
 import techcourse.myblog.service.exception.*;
 import techcourse.myblog.web.exception.NotLoggedInException;
